chore(ConvertInputTo): rename file and class
